Exception Handling

+++ src/main/java/com/siva/payrollmanagementsystem/repository/EmployeeRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.siva.payrollmanagementsystem.entity.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
+public interface EmployeeRepository extends JpaRepository<Employee, Long>{
 
+	
 }
